Conflicts:
	src/main/generated-sources/sarl/fr/utbm/ia51/agents/.Person.sarlbin
	src/main/generated-sources/sarl/fr/utbm/ia51/environment/.Environment.sarlbin

+++ src/main/java/fr/utbm/ia51/graph/environment/Courtyard.java
@@ -10,7 +10,7 @@ public class Courtyard extends StackPane{
 	private Rectangle building;
 		
 	
-	public Courtyard(int width, int height) {
+	public Courtyard(double width, double height) {
 		super();
         
         building= new Rectangle(width, height, Color.LIGHTGREY);
